chore: Update .gitignore and user model
This commit updates the .gitignore file to exclude the .env file from version control. It also adds new fields to the user model in the user.js file: username, password, phone, stripeId, and type. These changes were made to improve security and enhance the user management functionality of the application.