chore: Update angular.json and app-routing.module.ts
This commit updates the angular.json file to include the "analytics" property with a value of false in the "cli" section. It also adds a new route for the "register" page in the app-routing.module.ts file. These changes were made to disable analytics and add a new page for user registration.